LPD-25046 Move ScriptManagementConfigurationTestRule to test/rule
