+++ modules/apps/portal-security/portal-security-script-management-test-util/bnd.bnd
@@ -1,4 +1,6 @@
 Bundle-Name: Liferay Portal Security Script Management Test Utilities
 Bundle-SymbolicName: com.liferay.portal.security.script.management.test.util
 Bundle-Version: 1.1.1
-Export-Package: com.liferay.portal.security.script.management.test.util
+Export-Package:\
+	com.liferay.portal.security.script.management.test.rule,\
+	com.liferay.portal.security.script.management.test.util
