+++ modules/apps/portal-security/portal-security-script-management-test-util/src/main/java/com/liferay/portal/security/script/management/test/rule/ScriptManagementConfigurationTestRule.java
@@ -3,9 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.security.script.management.test.util;
+package com.liferay.portal.security.script.management.test.rule;
 
 import com.liferay.portal.kernel.test.rule.ClassTestRule;
+import com.liferay.portal.security.script.management.test.util.ScriptManagementConfigurationTestUtil;
 
 import org.junit.runner.Description;
