+++ modules/apps/portal-security/portal-security-script-management-test-util/src/main/resources/com/liferay/portal/security/script/management/test/rule/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
